// src/ts/enums/Hira.ts
export enum Hira {
	A = "あ",
	E = "え",
	I = "い",
	O = "お",
	U = "う",
	SA = "さ",
	SE = "せ",
	SHI = "し",
	SO = "そ",
	SU = "す",
	KA = "か",
	KE = "け",
	KI = "き",
	KO = "こ",
	KU = "く",
	TA = "た",
	TE = "て",
	CHI = "ち",
	TO = "と",
	TSU = "つ",
	N = "ん",
	NA = "な",
	NE = "ね",
	NI = "に",
	NO = "の",
	NU = "ぬ",
	HA = "は",
	HE = "へ",
	HI = "ひ",
	HO = "ほ",
	FU = "ふ",
	MA = "ま",
	ME = "め",
	MI = "み",
	MO = "も",
	MU = "む",
	RA = "ら",
	RE = "れ",
	RI = "り",
	RO = "ろ",
	RU = "る",
	YA = "や",
	YO = "よ",
	YU = "ゆ",
	WA = "わ",
	WO = "を",
}