// src/ts/enums/Kata.ts
export enum Kata {
	A = "ア",
	E = "エ",
	I = "イ",
	O = "オ",
	U = "ウ",

	SA = "サ",
	SE = "セ",
	SHI = "シ",
	SO = "ソ",
	SU = "ス",

	KA = "カ",
	KE = "ケ",
	KI = "キ",
	KO = "コ",
	KU = "ク",

	TA = "タ",
	TE = "テ",
	CHI = "チ",
	TO = "ト",
	TSU = "ツ",

	N = "ン",
	NA = "ナ",
	NE = "ネ",
	NI = "ニ",
	NO = "ノ",
	NU = "ヌ",

	HA = "ハ",
	HE = "ヘ",
	HI = "ヒ",
	HO = "ホ",
	FU = "フ",

	MA = "マ",
	ME = "メ",
	MI = "ミ",
	MO = "モ",
	MU = "ム",

	RA = "ラ",
	RE = "レ",
	RI = "リ",
	RO = "ロ",
	RU = "ル",

	YA = "ヤ",
	YO = "ヨ",
	YU = "ユ",

	WA = "ワ",
	WE = "ヱ",
	WI = "ヰ",
	WO = "ヲ",
}